Make Python 2 and explicit dependency
All of these scripts appear to be Python 2 specific and are not portable
to Python 3. The thing is Python 3 is often the default python
executable on many systems these days and will be more so going forward.
Using `#!/usr/bin/env python` will use the default system version making
these un-usable and un-testable on other systems.

The vast majority of systems that still have older Python versions as
the default also have it available under an explicit python2 name. This
includes CentOS as found on the test.door43.org server. Specifying the
shebang as `#!/usr/bin/env python2` will have no effect on systems where
that is the default while enabling all scripts to be usable on systems
where v3 is the default.