Port korn-shell scripts to full blown bash
Most of the syntax works fine anyway, and some of the bits that are a
bit kludgey to pull off in Korn will be a bit easier in Bash. This will
also make it easier to clean up the scripts for easier maintenance and
reading by the next guy along.