Refactor TeX export to use snippets local to repo
The assorted TeX snippets that were sitting in an httpdocs folder have
been added as part of this repository. Any changes to the latex
templating and processing will now be versioned along with the tools
that process them. The actual strings being placed are (or should be)
still retrieved from the API, but the code that lays out the document is
now referenced locally to the tools repo.

This also enables other people cloning the repo to adjust the layout to
their preferences as they generate output.