Reduce license page font size to it on one page
This is adjusted fer the default page size and the English language
license text. The value set here in Python that gets placed in the
ConTeXt template is just a default. If other languages have shorter or
longer requirements they can be set in the JSON data file using the
`licsize` key.

Suggestion for future improvement: wrap the license page in a self
adjusting loop much like the body pages use to not run over.