Make spaces after ordinal numbers non-breaking
The BCV references often have ordinal numbers before (as part of) the
Book name. The spaces between these ordinals and the book name is never
a good place for a line break. This expression should catch most
variants used in Western languages. E.g:

   (en) 1 Thessalonians
   (en) 3rd John
   (tr) 1. Selanikliler

Note this will not work for more complex scripts. For that a _real_
reference parser is going to necessary. This whole section is a dreadful
hack to start with, this patch just goes with the flow.