Merge build_all_DPB.sh and publish_PDF.sh scripts
The function of these two scripts was heavily overlapped. When to use
which one was confusing and they pretty much had to pass eachother their
own flags. In addition they ran a lot of the same checks and did a lot
of the same functions.

This script is a complete replacement for both scripts. In general it
performs the same functions but the way it is called has been sanitized.
Run with `-h` for options. All values may be specified with command line
flags AND/OR environment variables may be used to override the defaults.

For example to build multiple languages both usages are valid:

    ./obs/book/pdf_export.sh -l en -l ru -l ar
    langs=(en ru ar) ./obs/book/pdf_export.sh

Additional features not present or incomplete in the original scripts
include:

* Multiple output locations. A copy of the final PDF is placed in each
  of the requested output locations. This could be a flat directory for
  drafts (tip: use with `-t $(date +%F%m%d)` to tag by date) or it could
  be an API location (including relevant variable replacements such as
  `-o '/var/www/vhosts/api.unfoldingword.org/httpdocs/obs/pdf/$lang'`).
* Reporting. Add `-r LOC` flag(s) with directories or email addresses to
  receive a packaged version of all the files generated during the build
  run as well as some reports on the TeX output. If the debug flag `-d`
  is also specified the ConTeXt run will include output from trackers.