Remove mkdir function from Python export script
I'm sure this was intended to make things easier by creating output
directories is they don't exist, but it isn't implemented correctly and
actually fails is some situations: notable if the target _does_ exist
but with different permissions than expected or on file systems such as
TMPFS. Having this in here makes it very difficult to automate the
results. For example on my system I am using a RAM disk as the building
area and this prohibits that.

I think the best approach is to have this script focus on producing the
output and let automation scripts take care of other things like
creating directories. This script should have the smarts to process and
input format into an output format but shouldn't be too dependent on
where these things come from or where they go to.