Fix variable expansion for output directories
Oops. In spite of advertising that output paths could have variables I
didn't actually process them that way. This should actually allow things
like:

    ./pdf_export -o '/path/to/api/$lang/' -l tr -l ru

…to install respective PDFs to /path/to/api/tr/ and /path/to/api/ru/.

Note the single quotes so that the variable does _not_ gen expanded by
your shell and is saved for the script to expand for each language in
the loop.