revert title formatting changes that are ignored
Noto fonts dont' have small cap forms. Adding a small caps declaration
to this chapter title formatting rule only increases the chances that it
will _not_ be typeset correctly. Also forcing lower case when you have
special casing rules in your chapter names. I was using \MakeUpperCase
to make up for the lack of small caps, but if you don't want capitals at
all the thing to do is pass it the string formatted the way you want it
in the first place.