curl instead of wget and fix redirection
For reasons noted before curl is preferable to wget for single file
pulls. It handles bad network connections (especially SSL) better and
defaults to the kind of options you want for this operation (including
outputing to stdout).

Also if you are testing that a file doesn't exist, the thing to to is write
to it not append to it (which only makes sense for existing files).