re-enable (and document!) scripting features
Seeing as how they were disabled it is apparent that the reason for of
usage of these commands was not apparent. I suspect the previous
committer just wanted to see the temp files to debug things but didn't
know how to properly skip the garbage collection that cleans them up.

I'm enabling them again here, but this time I'm documenting what they do
how to use them.