fix bug: stop report from killing script if no 'bad' items found
This worked fine for me in testing because the report always had errors
to show so the grep never failed, but with the input data being cleaner
now the report has nothing to show. With `set -e` turned on (which is
important for other reasons) the greps in the report were killing the
script. This will only let them kill the report output section.

I also moved the report generation into the same BUILDDIR we are using
as scratch space to simplify the garbage collection when its all done.