fix portability: make sed usage BSD (OSX) compatible
The GNU and BSD seds process their arguments differently. Its possible
to write most commands in a way that works of both but you have to pay
attention. In the GNU side you can throw around the `-i` flag with or
without an argument for a backup extension anywhere you want and leave
off the `-e` and it will be assumed. On the BSD side the `-i` is much
greedier about looking for an argument. It will expect a zero-length
argument but the only way to make this happen without it consuming the
_next_ argument for itself (even if that is your expression) is to
explicitly use it before another flag, in most cases immediately
preceding the `-e` flag.