refactor 0f855e3: Added more descriptive variables
(Code changes by Richard, refactoring by Caleb)

So the changes in one commit came along after I had rebased everything
to clean up the merge issues. The contents of that commit reverted some
things I had fixed but changed others to make it easier for Richard to
use. What I've done here is take that commit diff line by line and apply
as many parts as I could without breaking what the extra features were
there for. I tried to use as much of the original variable naming as
possible while still keeping my other fixes.

The only real functional difference is in the way temp directories are
handled. If you specify a WORKING_DIR of your own, it will attempt to
use (or make if it doesn't exist) that. If not it will use the system
default temp dir plus a scratch space. It will only cleanup after itself
in this latter case (and only if DEBUG is false).

You can still reuse the automatically generated working directory if you
like, you just have to set it as the WORKING_DIR before the next run.