Typeset vertical bar (|) in a TeX compatible way
Fixes issue unfoldingWord-dev/Door43#245

Code point `U+007c` (|, also known as a pipe symbol) has special meaning
in most TeX variants and a literal vertical bar must be typeset with a
special command. The RegEx for finding and replacing these was broken.

This fixes both the finding and the replacing of vertical bars so that
they are properly typeset in TeX. This includes a fix that makes sure
all possible lines of output are processed for DocuWiki formatting
rather than just the first line.

This does _not_ include fixes to properly typeset cases where a vertical
bar has been used in the data when a different symbol is meant. Notably
Devanagari based languages such as Surgujia use a bar-like character for
punctuation:

> The end of a sentence or half-verse may be marked with a dot known as
> a pūrṇa virām or a vertical line danda: ।. The end of a full verse may
> be marked with two vertical lines: ॥.

Somebody used the ASCII vertical bar symbol instead of the correct marks
(`U+0964` or `U+0965`). This commit only corrects the typesetting of
actual legitimate vertical bars. Fixing the data to use the correct
symbol is work for another repository…