Setup max_chapters flag for easier debugging
There was already a variable setup in the script that could limit the
number of chapters retrieved and typeset, but one had to edit the
export.py script to access it. This moves it to a flag that can be
called even from the PDF generator script to make it easier to test run
various languages without having to edit scripts.