Add flag to edit the raw TeX before typesetting
Sometimes while debugging the typesetting process it's helpful to be able
to ① see and review the final source code after templating is complete
and ② have the option to hand edit it to experiment with the result.
This adds an edit flag to the output script that stops to open the file
in the system default editor before proceeding.